+++ config.py
@@ -2,6 +2,6 @@ _DB_CONF = {
  'host':'127.0.0.1',
  'port':3306,
  'user':'root',
- 'passwd':'aesa123!',
+ 'passwd':'baseball',
  'db':'baseball'
 }
